Auto stash before rebase of "feature/development"
11.31.2019
Lift calculation corrected
Author: Team NoName
Description: Made volume its own function and have it converting centimeters to meters. The balloon_buoyancy function has been cleaned up and is now using the inner volume to calculate mass of helium and outer volume to calculate mass of mylar and subtracting the weights from the buoyant force of the balloon.